Created extensions for CCRequestMessaging and split its functionalities
