+++ CCLocation/CCRequestMessaging.swift
@@ -145,427 +145,6 @@ class CCRequestMessaging: NSObject {
     //    }
     //}
     
-    func processIosSettings (serverMessage:Messaging_ServerMessage, store: Store<LibraryState>){
-        //        DDLogVerbose("got ios settings message")
-        
-        if (serverMessage.hasIosSettings && !serverMessage.iosSettings.hasGeoSettings) {
-            DispatchQueue.main.async {store.dispatch(DisableBackgroundGEOAction())}
-            DispatchQueue.main.async {store.dispatch(DisableForegroundGEOAction())}
-            DispatchQueue.main.async {store.dispatch(IsSignificationLocationChangeAction(isSignificantLocationChangeMonitoringState: false))}
-            DispatchQueue.main.async {store.dispatch(DisableCurrrentGEOAction())}
-        }
-        
-        if (serverMessage.hasIosSettings && !serverMessage.iosSettings.hasBeaconSettings) {
-            DispatchQueue.main.async {store.dispatch(DisableCurrentiBeaconMonitoringAction())}
-            DispatchQueue.main.async {store.dispatch(DisableForegroundiBeaconAction())}
-            DispatchQueue.main.async {store.dispatch(DisableBackgroundiBeaconAction())}
-            DispatchQueue.main.async {store.dispatch(DisableCurrrentiBeaconAction())}
-        }
-        
-        if (serverMessage.hasIosSettings && !serverMessage.iosSettings.hasInertialSettings) {
-            DispatchQueue.main.async {store.dispatch(DisableInertialAction())}
-        }
-        
-        if (serverMessage.hasIosSettings && serverMessage.iosSettings.hasGeoSettings) {
-            let geoSettings = serverMessage.iosSettings.geoSettings
-            
-            if geoSettings.hasSignificantUpates {
-                
-                if geoSettings.significantUpates {
-                    DispatchQueue.main.async {store.dispatch(IsSignificationLocationChangeAction(isSignificantLocationChangeMonitoringState: true))}
-                } else {
-                    DispatchQueue.main.async {store.dispatch(IsSignificationLocationChangeAction(isSignificantLocationChangeMonitoringState: false))}
-                }
-            } else {
-                DispatchQueue.main.async {store.dispatch(IsSignificationLocationChangeAction(isSignificantLocationChangeMonitoringState: false))}
-            }
-            
-            if geoSettings.hasBackgroundGeo {
-                
-                var activityType: CLActivityType?
-                
-                var maxRuntime:UInt64?
-                var minOffTime:UInt64?
-                
-                var desiredAccuracy:Int32?
-                var distanceFilter:Int32?
-                var pausesUpdates:Bool?
-                
-                if geoSettings.backgroundGeo.hasActivityType{
-                    switch geoSettings.backgroundGeo.activityType {
-                    case Messaging_IosStandardGeoSettings.Activity.other:
-                        activityType = .other
-                    case Messaging_IosStandardGeoSettings.Activity.auto:
-                        activityType = .automotiveNavigation
-                    case Messaging_IosStandardGeoSettings.Activity.fitness:
-                        activityType = .fitness
-                    case Messaging_IosStandardGeoSettings.Activity.navigation:
-                        activityType = .otherNavigation
-                    }
-                }
-                
-                if geoSettings.backgroundGeo.hasMaxRunTime {
-                    if geoSettings.backgroundGeo.maxRunTime > 0 {
-                        maxRuntime = geoSettings.backgroundGeo.maxRunTime
-                    }
-                }
-                
-                if geoSettings.backgroundGeo.hasMinOffTime {
-                    if geoSettings.backgroundGeo.minOffTime > 0 {
-                        minOffTime = geoSettings.backgroundGeo.minOffTime
-                    }
-                }
-                
-                if geoSettings.backgroundGeo.hasDistanceFilter {
-                    distanceFilter = geoSettings.backgroundGeo.distanceFilter
-                }
-                
-                if geoSettings.backgroundGeo.hasDesiredAccuracy {
-                    desiredAccuracy = geoSettings.backgroundGeo.desiredAccuracy
-                }
-                
-                if geoSettings.backgroundGeo.hasPausesUpdates {
-                    pausesUpdates = geoSettings.backgroundGeo.pausesUpdates
-                }
-                
-                let enableBackgroundGEOAction = EnableBackgroundGEOAction(
-                    activityType: activityType,
-                    maxRuntime: maxRuntime,
-                    minOffTime: minOffTime,
-                    desiredAccuracy: desiredAccuracy,
-                    distanceFilter: distanceFilter,
-                    pausesUpdates: pausesUpdates
-                )
-                
-                DispatchQueue.main.async {store.dispatch(enableBackgroundGEOAction)}
-            } else {
-                DispatchQueue.main.async {store.dispatch(DisableBackgroundGEOAction())}
-            }
-            
-            if geoSettings.hasForegroundGeo {
-                
-                var activityType: CLActivityType?
-                
-                var maxRuntime:UInt64?
-                var minOffTime:UInt64?
-                
-                var desiredAccuracy:Int32?
-                var distanceFilter:Int32?
-                var pausesUpdates:Bool?
-                
-                if geoSettings.foregroundGeo.hasActivityType{
-                    switch geoSettings.foregroundGeo.activityType {
-                    case Messaging_IosStandardGeoSettings.Activity.other:
-                        activityType = .other
-                    case Messaging_IosStandardGeoSettings.Activity.auto:
-                        activityType = .automotiveNavigation
-                    case Messaging_IosStandardGeoSettings.Activity.fitness:
-                        activityType = .fitness
-                    case Messaging_IosStandardGeoSettings.Activity.navigation:
-                        activityType = .otherNavigation
-                    }
-                }
-                
-                if geoSettings.foregroundGeo.hasMaxRunTime {
-                    if geoSettings.foregroundGeo.maxRunTime > 0 {
-                        maxRuntime = geoSettings.foregroundGeo.maxRunTime
-                    }
-                }
-                
-                if geoSettings.foregroundGeo.hasMinOffTime {
-                    if geoSettings.foregroundGeo.minOffTime > 0 {
-                        minOffTime = geoSettings.foregroundGeo.minOffTime
-                    }
-                }
-                
-                if geoSettings.foregroundGeo.hasDistanceFilter {
-                    distanceFilter = geoSettings.foregroundGeo.distanceFilter
-                }
-                
-                if geoSettings.foregroundGeo.hasDesiredAccuracy {
-                    desiredAccuracy = geoSettings.foregroundGeo.desiredAccuracy
-                }
-                
-                if geoSettings.foregroundGeo.hasPausesUpdates {
-                    pausesUpdates = geoSettings.foregroundGeo.pausesUpdates
-                }
-                
-                let enableForegroundGEOAction = EnableForegroundGEOAction(
-                    activityType: activityType,
-                    maxRuntime: maxRuntime,
-                    minOffTime: minOffTime,
-                    desiredAccuracy: desiredAccuracy,
-                    distanceFilter: distanceFilter,
-                    pausesUpdates: pausesUpdates
-                )
-                
-                DispatchQueue.main.async {store.dispatch(enableForegroundGEOAction)}
-            } else {
-                DispatchQueue.main.async {store.dispatch(DisableForegroundGEOAction())}
-            }
-        }
-        
-        if (serverMessage.hasIosSettings && serverMessage.iosSettings.hasBeaconSettings){
-            
-            let beaconSettings = serverMessage.iosSettings.beaconSettings
-            
-            if beaconSettings.hasMonitoring {
-                let monitoringSettings = beaconSettings.monitoring
-                
-                var monitoringRegions: [CLBeaconRegion] = []
-                
-                for region in monitoringSettings.regions {
-                    if region.hasUuid {
-                        if region.hasMajor {
-                            if region.hasMinor{
-                                if let uuid = UUID(uuidString: region.uuid) {
-                                    monitoringRegions.append(CLBeaconRegion(proximityUUID: uuid , major: CLBeaconMajorValue(region.major), minor: CLBeaconMinorValue(region.minor), identifier: "CC \(region.uuid):\(region.major):\(region.minor)"))
-                                }
-                            } else {
-                                if let uuid = UUID(uuidString: region.uuid) {
-                                    monitoringRegions.append(CLBeaconRegion(proximityUUID: uuid , major: CLBeaconMajorValue(region.major), identifier: "CC \(region.uuid):\(region.major)"))
-                                }
-                            }
-                        }
-                        else {
-                            if let uuid = UUID(uuidString: region.uuid) {
-                                monitoringRegions.append(CLBeaconRegion(proximityUUID: uuid, identifier: "CC \(region.uuid)"))
-                            }
-                        }
-                    }
-                }
-                
-                DispatchQueue.main.async {store.dispatch(EnableCurrentiBeaconMonitoringAction(monitoringRegions: monitoringRegions.sorted(by: {$0.identifier < $1.identifier})))}
-            } else {
-                DispatchQueue.main.async {store.dispatch(DisableCurrentiBeaconMonitoringAction())}
-            }
-            
-            if beaconSettings.hasForegroundRanging {
-                let foregroundRanging = beaconSettings.foregroundRanging
-                
-                var excludeRegions: [CLBeaconRegion] = []
-                var rangingRegions: [CLBeaconRegion] = []
-                
-                var maxRuntime:UInt64?
-                var minOffTime:UInt64?
-                var filterWindowSize:UInt64?
-                var maxObservations:UInt32?
-                
-                var eddystoneScan:Bool?
-                
-                for region in foregroundRanging.regions {
-                    if region.hasUuid {
-                        if region.hasMajor {
-                            if region.hasMinor{
-                                if let uuid = UUID(uuidString: region.uuid) {
-                                    rangingRegions.append(CLBeaconRegion(proximityUUID: uuid , major: CLBeaconMajorValue(region.major), minor: CLBeaconMinorValue(region.minor), identifier: "CC \(region.uuid):\(region.major):\(region.minor)"))
-                                }
-                            } else {
-                                if let uuid = UUID(uuidString: region.uuid) {
-                                    rangingRegions.append(CLBeaconRegion(proximityUUID: uuid , major: CLBeaconMajorValue(region.major), identifier: "CC \(region.uuid):\(region.major)"))
-                                }
-                            }
-                        }
-                        else {
-                            if let uuid = UUID(uuidString: region.uuid) {
-                                rangingRegions.append(CLBeaconRegion(proximityUUID: uuid, identifier: "CC \(region.uuid)"))
-                            }
-                        }
-                    }
-                }
-                
-                
-                for region in foregroundRanging.filter.excludeRegions {
-                    if region.hasUuid {
-                        if region.hasMajor {
-                            if region.hasMinor{
-                                if let uuid = UUID(uuidString: region.uuid) {
-                                    excludeRegions.append(CLBeaconRegion(proximityUUID: uuid , major: CLBeaconMajorValue(region.major), minor: CLBeaconMinorValue(region.minor), identifier: "CC \(region.uuid):\(region.major):\(region.minor)"))
-                                }
-                            } else {
-                                if let uuid = UUID(uuidString: region.uuid) {
-                                    excludeRegions.append(CLBeaconRegion(proximityUUID: uuid , major: CLBeaconMajorValue(region.major), identifier: "CC \(region.uuid):\(region.major)"))
-                                }
-                            }
-                        }
-                        else {
-                            if let uuid = UUID(uuidString: region.uuid) {
-                                excludeRegions.append(CLBeaconRegion(proximityUUID: uuid, identifier: "CC \(region.uuid)"))
-                            }
-                        }
-                    }
-                }
-                
-                if foregroundRanging.hasMaxRunTime {
-                    if foregroundRanging.maxRunTime > 0 {
-                        maxRuntime = foregroundRanging.maxRunTime
-                    }
-                }
-                
-                if foregroundRanging.hasMinOffTime {
-                    if foregroundRanging.minOffTime > 0 {
-                        minOffTime = foregroundRanging.minOffTime
-                    }
-                }
-                
-                if foregroundRanging.hasFilter {
-                    let filter = foregroundRanging.filter
-                    
-                    if filter.hasWindowSize {
-                        if filter.windowSize > 0 {
-                            filterWindowSize = filter.windowSize
-                        }
-                    }
-                    
-                    if filter.hasMaxObservations {
-                        if filter.maxObservations > 0 {
-                            maxObservations = filter.maxObservations
-                        }
-                    }
-                }
-                
-                if foregroundRanging.hasEddystoneScan {
-                    eddystoneScan = foregroundRanging.eddystoneScan
-                }
-                
-                let isIBeaconRangingEnabled = rangingRegions.count > 0 ? true : false
-                
-                DispatchQueue.main.async {store.dispatch(EnableForegroundBeaconAction(maxRuntime: maxRuntime,
-                                                                                      minOffTime: minOffTime,
-                                                                                      regions: rangingRegions.sorted(by: {$0.identifier < $1.identifier}),
-                                                                                      filterWindowSize: filterWindowSize,
-                                                                                      filterMaxObservations: maxObservations,
-                                                                                      filterExcludeRegions: excludeRegions.sorted(by: {$0.identifier < $1.identifier}),
-                                                                                      isEddystoneScanEnabled: eddystoneScan,
-                                                                                      isIBeaconRangingEnabled: isIBeaconRangingEnabled))}
-            } else {
-                DispatchQueue.main.async {store.dispatch(DisableForegroundiBeaconAction())}
-            }
-            
-            if beaconSettings.hasBackgroundRanging {
-                let backgroundRanging = beaconSettings.backgroundRanging
-                
-                var excludeRegions: [CLBeaconRegion] = []
-                var rangingRegions: [CLBeaconRegion] = []
-                
-                var maxRuntime:UInt64?
-                var minOffTime:UInt64?
-                var filterWindowSize:UInt64?
-                var maxObservations:UInt32?
-                
-                var eddystoneScan: Bool?
-                
-                for region in backgroundRanging.regions {
-                    if region.hasUuid {
-                        if region.hasMajor {
-                            if region.hasMinor{
-                                if let uuid = UUID(uuidString: region.uuid) {
-                                    rangingRegions.append(CLBeaconRegion(proximityUUID: uuid , major: CLBeaconMajorValue(region.major), minor: CLBeaconMinorValue(region.minor), identifier: "CC \(region.uuid):\(region.major):\(region.minor)"))
-                                }
-                            } else {
-                                if let uuid = UUID(uuidString: region.uuid) {
-                                    rangingRegions.append(CLBeaconRegion(proximityUUID: uuid , major: CLBeaconMajorValue(region.major), identifier: "CC \(region.uuid):\(region.major)"))
-                                }
-                            }
-                        }
-                        else {
-                            if let uuid = UUID(uuidString: region.uuid) {
-                                rangingRegions.append(CLBeaconRegion(proximityUUID: uuid, identifier: "CC \(region.uuid)"))
-                            }
-                        }
-                    }
-                }
-                
-                for region in backgroundRanging.filter.excludeRegions {
-                    if region.hasUuid {
-                        if region.hasMajor {
-                            if region.hasMinor{
-                                if let uuid = UUID(uuidString: region.uuid) {
-                                    excludeRegions.append(CLBeaconRegion(proximityUUID: uuid , major: CLBeaconMajorValue(region.major), minor: CLBeaconMinorValue(region.minor), identifier: "CC \(region.uuid):\(region.major):\(region.minor)"))
-                                }
-                            } else {
-                                if let uuid = UUID(uuidString: region.uuid) {
-                                    excludeRegions.append(CLBeaconRegion(proximityUUID: uuid , major: CLBeaconMajorValue(region.major), identifier: "CC \(region.uuid):\(region.major)"))
-                                }
-                            }
-                        }
-                        else {
-                            if let uuid = UUID(uuidString: region.uuid) {
-                                excludeRegions.append(CLBeaconRegion(proximityUUID: uuid, identifier: "CC \(region.uuid)"))
-                            }
-                        }
-                    }
-                }
-                
-                
-                if backgroundRanging.hasMaxRunTime {
-                    if backgroundRanging.maxRunTime > 0 {
-                        maxRuntime = backgroundRanging.maxRunTime
-                    }
-                }
-                
-                if backgroundRanging.hasMinOffTime {
-                    if backgroundRanging.minOffTime > 0 {
-                        minOffTime = backgroundRanging.minOffTime
-                    }
-                }
-                
-                if backgroundRanging.hasFilter {
-                    let filter = backgroundRanging.filter
-                    
-                    if filter.hasWindowSize {
-                        if filter.windowSize > 0 {
-                            filterWindowSize = filter.windowSize
-                        }
-                    }
-                    
-                    if filter.hasMaxObservations {
-                        if filter.maxObservations > 0 {
-                            maxObservations = filter.maxObservations
-                        }
-                        
-                    }
-                }
-                
-                if backgroundRanging.hasEddystoneScan {
-                    eddystoneScan = backgroundRanging.eddystoneScan
-                }
-                
-                let isIBeaconRangingEnabled = rangingRegions.count > 0 ? true : false
-                
-                DispatchQueue.main.async {self.stateStore.dispatch(EnableBackgroundiBeaconAction(maxRuntime: maxRuntime,
-                                                                                                 minOffTime: minOffTime,
-                                                                                                 regions: rangingRegions.sorted(by: {$0.identifier < $1.identifier}),
-                                                                                                 filterWindowSize: filterWindowSize,
-                                                                                                 filterMaxObservations: maxObservations,
-                                                                                                 filterExcludeRegions: excludeRegions.sorted(by: {$0.identifier < $1.identifier}),
-                                                                                                 eddystoneScanEnabled: eddystoneScan,
-                                                                                                 isIBeaconRangingEnabled: isIBeaconRangingEnabled))}
-            } else {
-                DispatchQueue.main.async {self.stateStore.dispatch(DisableBackgroundiBeaconAction())}
-            }
-        }
-        
-        if (serverMessage.hasIosSettings && serverMessage.iosSettings.hasInertialSettings) {
-            
-            var isInertialEnable: Bool?
-            var interval: UInt32?
-            
-            let inertialSettings = serverMessage.iosSettings.inertialSettings
-            
-            if inertialSettings.hasEnabled {
-                isInertialEnable = inertialSettings.enabled
-            }
-            
-            if inertialSettings.hasInterval {
-                interval = inertialSettings.interval
-            }
-            
-            DispatchQueue.main.async {self.stateStore.dispatch(InertialStateChangedAction(isEnabled: isInertialEnable,
-                                                                                          interval: interval))}
-        }
-    }
-    
     //    func processSystemBeacons(serverMessage:Messaging_ServerMessage) {
     //
     //        if (serverMessage.beacon.count > 0) {
@@ -607,81 +186,7 @@ class CCRequestMessaging: NSObject {
     //            }
     //        }
     //    }
-    
-    // MARK: - EVENT PROCESSING FUNCTIONS
-    
-    public func processIBeaconEvent(uuid:UUID, major:Int, minor:Int, rssi:Int, accuracy:Double, proximity:Int, timestamp:TimeInterval){
-        
-        let uuidData = uuid.uuidString.data(using: .utf8)
-        
-        var clientMessage = Messaging_ClientMessage()
-        var iBeaconMessage = Messaging_IBeacon()
-        
-        iBeaconMessage.uuid = uuidData!
-        iBeaconMessage.major = UInt32(major)
-        iBeaconMessage.minor = UInt32(minor)
-        iBeaconMessage.rssi = Int32(rssi)
-        iBeaconMessage.accuracy = accuracy
-        
-        iBeaconMessage.timestamp = UInt64(timestamp * 1000)
-        
-        iBeaconMessage.proximity = UInt32(proximity)
-        
-        clientMessage.ibeaconMessage.append(iBeaconMessage)
-        
-        //        Log.debug("iBeacon message built: \(clientMessage)")
-        
-        if let data = try? clientMessage.serializedData(){
-            sendOrQueueClientMessage(data: data, messageType: .queueable)
-        }
-    }
-    
-    public func processEddystoneEvent(eid:Data, tx:Int, rssi:Int, timestamp:TimeInterval){
-        
-        var clientMessage = Messaging_ClientMessage()
-        var eddyStoneMessage = Messaging_EddystoneBeacon()
-        
-        eddyStoneMessage.eid = eid
-        eddyStoneMessage.rssi = Int32(rssi)
-        eddyStoneMessage.timestamp = UInt64(timestamp * 1000)
-        eddyStoneMessage.tx = Int32(tx)
-        
-        clientMessage.eddystonemessage.append(eddyStoneMessage)
-        
-        Log.verbose("Eddystone beacon message build: \(clientMessage)")
-        
-        if let data = try? clientMessage.serializedData(){
-            sendOrQueueClientMessage(data: data, messageType: .queueable)
-        }
-    }
-    
-    public func processBluetoothEvent(uuid:UUID, rssi:Int, timeInterval:TimeInterval) {
-        
-        let uuidData = uuid.uuidString.data(using: .utf8)
-        
-        var clientMessage = Messaging_ClientMessage()
-        var bluetoothMessage = Messaging_Bluetooth()
-        
-        //        var uuidBytes: [UInt8] = [UInt8](repeating: 0, count: 16)
-        //        uuid.getBytes(&uuidBytes)
-        //        let uuidData = NSData(bytes: &uuidBytes, length: 16)
-        
-        bluetoothMessage.identifier = uuidData!
-        bluetoothMessage.rssi = Int32(rssi)
-        bluetoothMessage.tx = 0
-        bluetoothMessage.timestamp = UInt64(fabs(timeInterval * Double(1000.0)))
-        
-        clientMessage.bluetoothMessage.append(bluetoothMessage)
-        
-        //DDLogVerbose ("Bluetooth message build: \(clientMessage)")
-        
-        if let data = try? clientMessage.serializedData() {
-            sendOrQueueClientMessage(data: data, messageType: .queueable)
-        }
-    }
-    
-    
-    
+
     //    func sendCollatedBluetoothMessage(devices:Dictionary<String, Dictionary<String, Int>>, timeInterval:TimeInterval) {
     //
     //        var clientMessage = Messaging_ClientMessage()
@@ -711,92 +216,6 @@ class CCRequestMessaging: NSObject {
     //            sendClientMessage(data: data, messageType: .queueable)
     //        }
     //    }
-    
-    public func processLocationEvent(location:CLLocation) {
-        
-        let userDefaults = UserDefaults.standard
-        
-        var clientMessage = Messaging_ClientMessage()
-        var locationMessage = Messaging_LocationMessage()
-        
-        var counter = userDefaults.integer(forKey: CCRequestMessagingConstants.messageCounter)
-        
-        if counter < Int.max {
-            counter = counter + 1
-        } else {
-            counter = 0
-        }
-        
-        locationMessage.longitude = location.coordinate.longitude
-        locationMessage.latitude = location.coordinate.latitude
-        locationMessage.horizontalAccuracy = location.horizontalAccuracy
-        locationMessage.verticalAccuracy = location.verticalAccuracy
-        locationMessage.course = Double(counter)
-        locationMessage.speed = 1
-        
-        // a negative value for vertical accuracy indicates that the altitude value is invalid
-        if (location.verticalAccuracy >= 0){
-            locationMessage.altitude = location.altitude
-        }
-        
-        let trueTimeSame = timeHandling.isRebootTimeSame(stateStore: stateStore, ccSocket: ccSocket)
-        
-        if ((stateStore.state.ccRequestMessagingState.libraryTimeState?.lastTrueTime) != nil || trueTimeSame) {
-            
-            let lastSystemTime = stateStore.state.ccRequestMessagingState.libraryTimeState?.systemTimeAtLastTrueTime
-            
-            let currentTime = Date()
-            
-            let beetweenSystemsTimeInterval = currentTime.timeIntervalSince(lastSystemTime!)
-            
-            let sendTimeInterval = stateStore.state.ccRequestMessagingState.libraryTimeState?.lastTrueTime?.addingTimeInterval(beetweenSystemsTimeInterval).timeIntervalSince1970
-            
-            locationMessage.timestamp = UInt64(sendTimeInterval! * 1000)
-            
-            if !trueTimeSame {
-                locationMessage.speed = -1
-            }
-        } else {
-            locationMessage.timestamp = UInt64(0)
-        }
-        
-        clientMessage.locationMessage.append(locationMessage)
-        
-        if let data = try? clientMessage.serializedData(){
-            //            NSLog("Location message build: \(clientMessage) with size: \(String(describing: data.count))")
-            userDefaults.set(counter, forKey: CCRequestMessagingConstants.messageCounter)
-            sendOrQueueClientMessage(data: data, messageType: .queueable)
-        }
-    }
-    
-    public func processStep(date: Date, angle: Double) {
-        
-        var clientMessage = Messaging_ClientMessage()
-        var stepMessage = Messaging_Step()
-        
-        stepMessage.angle = angle
-        
-        let trueTimeSame = timeHandling.isRebootTimeSame(stateStore: stateStore, ccSocket: ccSocket)
-        
-        if ((stateStore.state.ccRequestMessagingState.libraryTimeState?.lastTrueTime) != nil || trueTimeSame) {
-            
-            let lastSystemTime = stateStore.state.ccRequestMessagingState.libraryTimeState?.systemTimeAtLastTrueTime
-            
-            let beetweenSystemsTimeInterval = date.timeIntervalSince(lastSystemTime!)
-            
-            let sendTimeInterval = stateStore.state.ccRequestMessagingState.libraryTimeState?.lastTrueTime?.addingTimeInterval(beetweenSystemsTimeInterval).timeIntervalSince1970
-            
-            stepMessage.timestamp = UInt64(sendTimeInterval! * 1000)
-        }
-        
-        clientMessage.step.append(stepMessage)
-        
-        if let data = try? clientMessage.serializedData(){
-//            NSLog("Step message build: \(clientMessage) with size: \(String(describing: data.count))")
-            sendOrQueueClientMessage(data: data, messageType: .queueable)
-        }
-    }
-    
     
     public func processIOSCapability(locationAuthStatus: CLAuthorizationStatus?,
                                      bluetoothHardware: CBCentralManagerState?,
@@ -869,47 +288,6 @@ class CCRequestMessaging: NSObject {
         }
     }
     
-    public func processAliases(aliases:Dictionary<String,String>) {
-        
-        var clientMessage = Messaging_ClientMessage()
-        
-        for key in aliases.keys{
-            
-            var aliasMessage = Messaging_AliasMessage()
-            
-            aliasMessage.key = key
-            aliasMessage.value = aliases[key]!
-            
-            clientMessage.alias.append(aliasMessage)
-        }
-        
-        //DDLogVerbose ("alias message build: \(clientMessage)")
-        
-        if let data = try? clientMessage.serializedData() {
-            sendOrQueueClientMessage(data: data, messageType: .discardable)
-        }
-    }
-    
-    public func processMarker(data:String) {
-        
-        if let timeInterval = TimeHandling.getCurrentTimePeriodSince1970(stateStore: stateStore) {
-            
-            var clientMessage = Messaging_ClientMessage()
-            var markerMessage = Messaging_MarkerMessage()
-            
-            markerMessage.data = data
-            markerMessage.time = UInt64(fabs(timeInterval * Double(1000.0)))
-            
-            clientMessage.marker = markerMessage
-            
-            //DDLogVerbose ("marker message build: \(clientMessage)")
-            
-            if let data = try? clientMessage.serializedData() {
-                sendOrQueueClientMessage(data: data, messageType: .queueable)
-            }
-        }
-    }
-    
     //- (void) sendAcknowledgement:(NSString*)messageId {
     //
     //
@@ -930,7 +308,6 @@ class CCRequestMessaging: NSObject {
     //
     //    [self sendClientMessage:[self getDataForClientMessage:clientMessage] queuable:TRUE];
     //}
-    
     
     // MARK: - STATE HANDLING FUNCTIONS
     
@@ -1021,158 +398,6 @@ class CCRequestMessaging: NSObject {
         DispatchQueue.main.async {self.stateStore.dispatch(ScheduleSilencePeriodTimerAction())}
     }
     
-    public func sendQueuedClientMessages(firstMessage: Data?) {
-        
-        var workItem: DispatchWorkItem!
-        
-        if (firstMessage != nil){
-            
-            // If radioSilencer is nil, message is sent to the server instant
-            if stateStore.state.ccRequestMessagingState.radiosilenceTimerState?.timeInterval == nil,
-                let newMessage = firstMessage {
-                
-                sendMessageDirectly(message: newMessage)
-                return
-                
-            } else {
-                if let newMessage = firstMessage {
-                    insertMessageInLocalDatabase(message: newMessage)
-                }
-            }
-
-        }
-        
-        workItem = DispatchWorkItem { [weak self] in
-            
-            if !workItem.isCancelled {
-                
-                if self == nil {
-                    //This should not happen
-                    Log.error("Self object is nil inside sending message method. Message won't be delivered or queued")
-                    return
-                }
-                
-                let maxMessagesToReturn = 100
-                
-                var connectionState = self?.stateStore.state.ccRequestMessagingState.webSocketState?.connectionState
-                
-                while (self?.getMessageCount() ?? -1 > 0 && connectionState == .online) {
-                    
-                    if workItem.isCancelled { break }
-                    
-                    connectionState = self?.stateStore.state.ccRequestMessagingState.webSocketState?.connectionState
-                    
-                    var compiledClientMessage = Messaging_ClientMessage()
-                    var backToQueueMessages = Messaging_ClientMessage()
-                    
-                    var tempMessageData:[Data]?
-                    var subMessageCounter:Int = 0
-                    
-                    let messageNumber = self?.getMessageCount() ?? -1
-                    
-                    if (messageNumber == 0) {
-                        Log.verbose ("No queued messages available to send")
-                    }
-                    
-                    Log.verbose ("\(messageNumber) Queued messages are available")
-                    
-                    while (self?.getMessageCount() ?? -1 > 0 && subMessageCounter < maxMessagesToReturn) {
-                        
-                        if workItem.isCancelled { break }
-                        
-                        tempMessageData = self?.popMessagesFromLocalDatabase(maxMessagesToReturn: maxMessagesToReturn)
-                        
-                        if let unwrappedTempMessageData = tempMessageData {
-                            for tempMessage in unwrappedTempMessageData {
-                                
-                                if workItem.isCancelled { break }
-                                
-                                let (newSubMessageCounter, newCompiledClientMessage, newBackToQueueMessages) = self?.handleMessageType(message: tempMessage,
-                                                                                                                 subMessageInitialNumber: subMessageCounter,
-                                                                                                                 compiledMessage: compiledClientMessage,
-                                                                                                                 queueMessage: backToQueueMessages) ?? (subMessageCounter, Messaging_ClientMessage(),Messaging_ClientMessage())
-                                subMessageCounter = newSubMessageCounter
-                                compiledClientMessage = newCompiledClientMessage
-                                backToQueueMessages = newBackToQueueMessages
-                            }
-                        }
-                    }
-                    
-                    //DDLogVerbose("Compiled \(subMessageCounter) message(s)")
-                    
-                    //        if (compiledClientMessage.locationMessage.count > 0){
-                    //            let geoMsg = compiledClientMessage.locationMessage[0]
-                    //            let geoData = try? geoMsg.serializedData()
-                    //DDLogVerbose("compiled geoMsg: \(geoData?.count ?? -1) and byte array: \(geoData?.hexEncodedString() ?? "NOT AVAILABLE")")
-                    //        }
-                    
-                    //        if (compiledClientMessage.bluetoothMessage.count > 0){
-                    //            let blMsg = compiledClientMessage.bluetoothMessage[0]
-                    //            let blData = try? blMsg.serializedData()
-                    //DDLogVerbose("compiled bluetooth message: \(blData?.count ?? -1) and byte array: \(blData?.hexEncodedString() ?? "NOT AVAILABLE"))")
-                    
-                    //        }
-                    
-                    //        for beacon in compiledClientMessage.ibeaconMessage {
-                    //            DDLogVerbose("Sending beacons \(compiledClientMessage.ibeaconMessage.count) with \(beacon)")
-                    //        }
-                    
-                    //        if (compiledClientMessage.alias.count > 0){
-                    //            let alMsg = compiledClientMessage.alias[0]
-                    //            let alData = try? alMsg.serializedData()
-                    //DDLogVerbose("compiled alias message: \(alData?.count ?? -1)  and byte array: \(alData?.hexEncodedString() ?? "NOT AVAILABLE"))")
-                    //        }
-                    
-                    if workItem.isCancelled { break }
-                    
-                    if let backToQueueData = try? backToQueueMessages.serializedData() {
-                        //DDLogDebug("Had to split a client message into two, pushing \(subMessageCounter) unsent messages back to the Queue")
-                        if backToQueueData.count > 0 {
-                            //    ccRequest?.messageQueuePushSwiftBridge(backToQueueData)
-                            
-                            self?.insertMessageInLocalDatabase(message: backToQueueData)
-                        }
-                    } else {
-                        //DDLogError("Couldn't serialize back to queue data")
-                    }
-                    
-                    if let data = try? compiledClientMessage.serializedData(){
-                        if (data.count > 0) {
-                            
-                            if let stateStore = self?.stateStore {
-                                if let ccSocket = self?.ccSocket {
-                                    if let isRebootTimeSame = self?.timeHandling.isRebootTimeSame(stateStore: stateStore, ccSocket: ccSocket) {
-                                        if isRebootTimeSame {
-                                            if let currentTimePeriod = TimeHandling.getCurrentTimePeriodSince1970(stateStore: stateStore) {
-                                                compiledClientMessage.sentTime = UInt64(currentTimePeriod * 1000)
-                                                //                        DDLogVerbose("Added sent time to the client message")
-                                            }
-                                        }
-                                    }
-                                    
-                                    if let dataIncludingSentTime = try? compiledClientMessage.serializedData(){
-                                        //            Log.verbose("Sending \(unwrappedData.count) bytes of compiled client message data")
-                                        self?.ccSocket?.sendWebSocketMessage(data: dataIncludingSentTime)
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-                
-                if let index = self?.workItems.index(where: {$0 === workItem!}) {
-                    self?.workItems.remove(at: index)
-                }
-                
-                workItem = nil
-            }
-        }
-        
-        workItems.append(workItem)
-        
-        DispatchQueue.global(qos: .background).async(execute: workItem)
-    }
-    
     // MARK: - APPLICATION STATE HANDLING FUNCTIONS
     
     @objc func applicationWillResignActive () {
@@ -1230,8 +455,6 @@ class CCRequestMessaging: NSObject {
                                                name:UIApplication.willTerminateNotification,
                                                object:nil)
     }
-    
-    
     
     func setupBatteryStateAndLevelNotifcations (){
         UIDevice.current.isBatteryMonitoringEnabled = true
@@ -1320,121 +543,6 @@ extension CCRequestMessaging: TimeHandlingDelegate {
                 if radioSilenceTimerState.startTimeInterval != nil {
                     DispatchQueue.main.async {self.stateStore.dispatch(ScheduleSilencePeriodTimerAction())}
                 }
-            }
-        }
-    }
-}
-
-// MARK:- StoreSubscriber delegate
-extension CCRequestMessaging: StoreSubscriber {
-    public func newState(state: CCRequestMessagingState) {
-//        DDLogDebug("new state is: \(state)")
-        
-        if let webSocketState = state.webSocketState {
-            if webSocketState != currentWebSocketState{
-                currentWebSocketState = webSocketState
-                
-                if webSocketState.connectionState == ConnectionState.online {
-                    
-                    let aliases: Dictionary? = UserDefaults.standard.dictionary(forKey: CCSocketConstants.ALIAS_KEY)
-                    
-                    if (aliases != nil){
-                        processAliases(aliases: aliases! as! Dictionary<String, String>)
-                    }
-                }
-            }
-        }
-        
-        // if we have a radioSilenceTimer
-        if let newTimerState = state.radiosilenceTimerState {
-            
-            // and if its state has changed
-            if newTimerState != currentRadioSilenceTimerState {
-                
-                currentRadioSilenceTimerState = newTimerState
-                
-                if newTimerState.timer == .schedule {
-                    
-                    if let timeInterval = newTimerState.timeInterval {
-//                        DDLogVerbose("RADIOSILENCETIMER trying to schedule timer with timeInterval = \(timeInterval / 1000)")
-                        
-                        if timeBetweenSendsTimer != nil {
-                            if timeBetweenSendsTimer!.isValid{
-                                timeBetweenSendsTimer!.invalidate()
-                            }
-                        }
-                        
-                        if let radioSilenceTimerState = newTimerState.startTimeInterval {
-                            
-                            let intervalForLastTimer = TimeHandling.timeIntervalSinceBoot() - radioSilenceTimerState
-//                            DDLogVerbose("RADIOSILENCETIMER intervalForLastTimer = \(intervalForLastTimer)")
-                            
-                            if intervalForLastTimer < Double(Double(timeInterval) / 1000) {
-                                timeBetweenSendsTimer = Timer.scheduledTimer(timeInterval:TimeInterval(intervalForLastTimer), target: self, selector: #selector(self.sendQueuedClientMessagesTimerFiredOnce), userInfo: nil, repeats: false)
-                                DispatchQueue.main.async {self.stateStore.dispatch(TimerRunningAction(startTimeInterval: nil))}
-                            } else {
-                                timeBetweenSendsTimer = Timer.scheduledTimer(timeInterval:TimeInterval(Double(timeInterval) / 1000), target: self, selector: #selector(self.sendQueuedClientMessagesTimerFired), userInfo: nil, repeats: true)
-                                
-                                DispatchQueue.main.async {self.stateStore.dispatch(TimerRunningAction(startTimeInterval: TimeHandling.timeIntervalSinceBoot()))}
-                            }
-                            
-                        } else {
-                            
-                            timeBetweenSendsTimer = Timer.scheduledTimer(timeInterval:TimeInterval(Double(timeInterval) / 1000), target: self, selector: #selector(self.sendQueuedClientMessagesTimerFired), userInfo: nil, repeats: true)
-                            
-                            DispatchQueue.main.async {self.stateStore.dispatch(TimerRunningAction(startTimeInterval: TimeHandling.timeIntervalSinceBoot()))}
-                        }
-                    }
-                }
-                
-                if newTimerState.timer == .running {
-//                    DDLogVerbose("RADIOSILENCETIMER timer is in running state")
-                }
-                
-//                covers case were app starts from terminated and no timer is available yet
-                if timeBetweenSendsTimer == nil {
-//                    DDLogVerbose("RADIOSILENCETIMER timeBetweenSendsTimer == nil, scheduling new timer")
-                    if timeHandling.isRebootTimeSame(stateStore: stateStore, ccSocket: ccSocket){
-                        DispatchQueue.main.async {self.stateStore.dispatch(ScheduleSilencePeriodTimerAction())}
-                    }
-                }
-                
-                if newTimerState.timer == .invalidate {
-//                    DDLogVerbose("RADIOSILENCETIMER invalidate timer")
-                    if timeBetweenSendsTimer != nil{
-                        if timeBetweenSendsTimer!.isValid {
-                            timeBetweenSendsTimer!.invalidate()
-                        }
-                        timeBetweenSendsTimer = nil
-                    }
-                    
-                    DispatchQueue.main.async {self.stateStore.dispatch(TimerStoppedAction())}
-                }
-            }
-        }
-        
-        if let newLibraryTimeState = state.libraryTimeState {
-            if newLibraryTimeState != currentLibraryTimerState {
-                
-                currentLibraryTimerState = newLibraryTimeState
-                
-                if let bootTimeInterval = newLibraryTimeState.bootTimeIntervalAtLastTrueTime {
-                    
-                    let timeDifferenceSinceLastTrueTime = bootTimeInterval - TimeHandling.timeIntervalSinceBoot()
-                    
-                    if timeDifferenceSinceLastTrueTime > 60 {
-                        timeHandling.fetchTrueTime()
-                    }
-                }
-            }
-        }
-        
-        if let newCapabilityState = state.capabilityState {
-            if newCapabilityState != currentCapabilityState {
-                
-                processIOSCapability(locationAuthStatus: newCapabilityState.locationAuthStatus, bluetoothHardware: newCapabilityState.bluetoothHardware, batteryState: newCapabilityState.batteryState, isLowPowerModeEnabled: newCapabilityState.isLowPowerModeEnabled, isLocationServicesEnabled: newCapabilityState.isLocationServicesAvailable)
-                
-                currentCapabilityState = newCapabilityState
             }
         }
     }
